+++ 2020-11/spring-09/spring-09-exercise/src/main/java/ru/otus/spring/dao/AuthorDao.java
@@ -1,4 +1,6 @@
-package ru.otus.spring.domain;
+package ru.otus.spring.dao;
+
+import ru.otus.spring.domain.Author;
 
 import java.util.Optional;
