+++ 2020-11/spring-09/spring-09-exercise/src/main/java/ru/otus/spring/dao/GenreDao.java
@@ -1,4 +1,6 @@
-package ru.otus.spring.domain;
+package ru.otus.spring.dao;
+
+import ru.otus.spring.domain.Genre;
 
 import java.util.Optional;
