+++ 2020-11/spring-09/spring-09-exercise/src/main/java/ru/otus/spring/dao/GenreJpa.java
@@ -1,22 +1,22 @@
-package ru.otus.spring.domain;
+package ru.otus.spring.dao;
 
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
+import ru.otus.spring.domain.Genre;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import java.util.Optional;
 
-
 @Transactional
 @Repository
-public class AuthorJpa implements AuthorDao {
+public class GenreJpa implements GenreDao {
     @PersistenceContext
     private EntityManager em;
 
     @Override
-    public Optional<Author> findById(long authorId) {
-        return Optional.ofNullable(em.find(Author.class, authorId));
+    public Optional<Genre> findById(long genreId) {
+        return Optional.ofNullable(em.find(Genre.class, genreId));
     }
 
 }
